Modify coordinate painting

+++ BigWinner/PaintFun.h
@@ -21,6 +21,6 @@ public:
 
 	void DrawRect(HDC *hdcBuffer, COLORREF color, RECT rect, char pen);
 	void DrawFrame(HDC *hdc, RECT rect);
-	void DrawCoordinate(HDC *hdc, int xlocation, int ylocation, int len, int cnt, int start, unsigned char coortype);
+	void DrawCoordinate(HDC *hdc, int xlocation, int ylocation, int len, int cnt, float start, unsigned char coortype);
 };
